refactor: Move wallet generation to utils
